% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltmix-package.R
\name{ltmix}
\alias{ltmix}
\alias{_PACKAGE}
\title{ltmix: Left-Truncated Mixtures of Gamma, Weibull, and Lognormal Distributions}
\description{
Mixture modelling of one-dimensional data using combinations of left-truncated Gamma,
Weibull, and Lognormal Distributions.
}
