% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{ltmm}
\alias{ltmm}
\title{Fit a Left-truncated mixture model (LTMM)}
\usage{
ltmm(x, G, distributions, trunc = NULL, EM_init_method = "emEM",
  EM_starts = 5, init_pars = NULL, init_pi = NULL,
  init_classes = NULL, one_group_reps = 50, eps = 1e-06,
  max.it = 1000, verbose = FALSE)
}
\arguments{
\item{x}{data vector}

\item{G}{number of components}

\item{distributions}{densities to combine}

\item{trunc}{left truncation point (optional)}

\item{EM_init_method}{initialization method for EM algorithm}

\item{EM_starts}{number of random starts for initialization of EM algorithm. (only for G > 1)}

\item{init_pars}{initial parameter values (list of length G)}

\item{init_pi}{manually specified initial component proportions (for init_method=specified)}

\item{init_classes}{manually specified initial classes. will overwrite init_pars and init_pi}

\item{one_group_reps}{number of random starts for each numerical optimization in 1-component model}

\item{eps}{stopping tolerance for EM algoithm}

\item{max.it}{maximum number of iterations of EM algorithm}

\item{verbose}{print information as fitting progresses?}
}
\value{
an ltmm model object
}
\description{
This function generate a mixture of lognormal, gamma, and weibull distributions
}
\details{
Data set x and number of components G are provided
}
