% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{secura}
\alias{secura}
\title{The Secura Belgian Re Data}
\format{
An object of class \code{data.frame} with 371 rows and 2 columns.
}
\usage{
secura
}
\description{
"The Secura Belgian Re data set contains automobile claims from 1988 until 2001, which are at least as large as 1,200,000 Euros." (Beirlant, Goegebeur, Segers & Teugels, 2004).
}
\references{
Beirlant, J., Goegebeur Y., Segers, J., & Teugels, J. Statistics of extremes : theory and applications. Hoboken, NJ: Wiley, 2004. Print.
}
\keyword{datasets}
