% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_level.R
\name{createLtmmObj}
\alias{createLtmmObj}
\title{Create an ltmm model object given data and parameters}
\usage{
createLtmmObj(x, distributions, trunc, Pars, Pi, npars = NULL)
}
\arguments{
\item{x}{data vector}

\item{distributions}{densities to combine}

\item{trunc}{left truncation point (optional)}

\item{Pars}{list of length G of parameter values}

\item{Pi}{vector of length G of component proportions}

\item{npars}{Can optionally be used to overwrite the number of free parameters (used in the
calculation of AIC & BIC), if the model has additional constraints}
}
\value{
An ltmm model object
}
\description{
This function is useful for omparing models produced using the ltmix package
              to models fit using other, or for computing fit criteria and risk measures for
              a known set of parameters.
}
